+++ Gruntfile.js
@@ -5,11 +5,10 @@ const path = require('path');
 const debug = require('debug');
 
 const webpackConfig = require('./build/webpack.config');
+const webpackConfigTest = require('./test/test.webpack.config');
 const pkg = require('./package.json');
-const hookRcedit = require('./build/util/hook-rcedit');
 const codeSignConfig = require('../keys/codesign');
 
-hookRcedit.setup();
 debug.enable('electron-notarize');
 
 module.exports = function(grunt) {
@@ -31,6 +30,14 @@ module.exports = function(grunt) {
     const zipCommentPlaceholder =
         zipCommentPlaceholderPart + '.'.repeat(512 - zipCommentPlaceholderPart.length);
     const electronVersion = pkg.dependencies.electron.replace(/^\D/, '');
+
+    const webpackOptions = {
+        date,
+        beta: !!grunt.option('beta'),
+        get sha() {
+            return grunt.config.get('gitinfo.local.branch.current.shortSHA');
+        }
+    };
 
     grunt.initConfig({
         gitinfo: {
@@ -214,11 +221,12 @@ module.exports = function(grunt) {
             }
         },
         webpack: {
-            js: webpackConfig.config(grunt)
+            app: webpackConfig.config(webpackOptions),
+            test: webpackConfigTest
         },
         'webpack-dev-server': {
             options: {
-                webpack: webpackConfig.config(grunt, 'development'),
+                webpack: webpackConfig.config({ ...webpackOptions, mode: 'development' }),
                 publicPath: '/',
                 contentBase: path.resolve(__dirname, 'tmp'),
                 progress: false
@@ -226,16 +234,6 @@ module.exports = function(grunt) {
             js: {
                 keepalive: true,
                 port: 8085
-            }
-        },
-        watch: {
-            options: {
-                interrupt: true,
-                debounceDelay: 500
-            },
-            indexhtml: {
-                files: 'app/index.html',
-                tasks: ['copy:html']
             }
         },
         electron: {
@@ -580,6 +578,12 @@ module.exports = function(grunt) {
                     ]
                 }
             }
+        },
+        'run-test': {
+            options: {
+                headless: true
+            },
+            default: 'test/runner.html'
         }
     });
 };