+++ app/scripts/comp/app/shortcuts.js
@@ -23,7 +23,9 @@ const globalShortcuts = {
     copyPassword: { mac: 'Ctrl+Alt+C', all: 'Shift+Alt+C' },
     copyUser: { mac: 'Ctrl+Alt+B', all: 'Shift+Alt+B' },
     copyUrl: { mac: 'Ctrl+Alt+U', all: 'Shift+Alt+U' },
-    autoType: { mac: 'Ctrl+Alt+T', all: 'Shift+Alt+T' }
+    copyOtp: {},
+    autoType: { mac: 'Ctrl+Alt+T', all: 'Shift+Alt+T' },
+    restoreApp: {}
 };
 
 const Shortcuts = {
@@ -48,6 +50,9 @@ const Shortcuts = {
         };
     },
     presentShortcut(shortcutValue, formatting) {
+        if (!shortcutValue) {
+            return '-';
+        }
         return shortcutValue
             .split(/\+/g)
             .map(part => {