+++ app/scripts/models/app-model.js
@@ -327,8 +327,22 @@ class AppModel {
 
     prepareFilter(filter) {
         filter = { ...filter };
+
         filter.textLower = filter.text ? filter.text.toLowerCase() : '';
+        filter.textParts = null;
+        filter.textLowerParts = null;
+
+        const exact = filter.advanced && filter.advanced.exact;
+        if (!exact && filter.text) {
+            const textParts = filter.text.split(/\s+/).filter(s => s);
+            if (textParts.length) {
+                filter.textParts = textParts;
+                filter.textLowerParts = filter.textLower.split(/\s+/).filter(s => s);
+            }
+        }
+
         filter.tagLower = filter.tag ? filter.tag.toLowerCase() : '';
+
         return filter;
     }
 
@@ -423,7 +437,7 @@ class AppModel {
     createNewFile(name) {
         if (!name) {
             for (let i = 0; ; i++) {
-                name = 'New' + (i || '');
+                name = Locale.openNewFile + (i || '');
                 if (!this.files.getByName(name) && !this.fileInfos.getByName(name)) {
                     break;
                 }
@@ -478,6 +492,9 @@ class AppModel {
                 params,
                 (err, file) => {
                     if (err) {
+                        if (err.name === 'KdbxError') {
+                            return callback(err);
+                        }
                         logger.info(
                             'Error loading file from cache, trying to open from storage',
                             err
@@ -501,6 +518,9 @@ class AppModel {
                         setTimeout(() => this.syncFile(file), 0);
                         callback(err);
                     } else {
+                        if (err.name === 'KdbxError') {
+                            return callback(err);
+                        }
                         logger.info(
                             'Error loading file from cache, trying to open from storage',
                             err