+++ app/scripts/views/footer-view.js
@@ -38,7 +38,7 @@ class FooterView extends View {
 
     render() {
         super.render({
-            files: this.model.files.toJSON(),
+            files: this.model.files,
             updateAvailable: ['ready', 'found'].indexOf(UpdateModel.updateStatus) >= 0
         });
     }