+++ test/index.js
@@ -1,0 +1,4 @@
+function requireAll(req) {
+    req.keys().forEach(req);
+}
+requireAll(require.context('test/src/', true, /\.js$/));